requirements.txt
yt-dlp>=2024.7.1
srt>=3.5.3
webvtt-py>=0.4.6
wordfreq>=3.1
regex>=2024.5.15
requests>=2.32.3
typer>=0.12.3
rich>=13.7.1
subliminal>=2.1.0
babelfish>=0.6.1
numpy>=1.26.0
